@@ -13,14 +13,15 @@
   },
   "dependencies": {
     "backbone": "^1.1.2",
+    "classnames": "^2.1.2",
     "flux": "^2.0.3",
     "jquery": "^2.1.3",
     "lazy.js": "^0.4.0",
     "moment": "^2.9.0",
     "mousetrap": "^1.4.6",
     "oboe": "^2.1.2",
-    "react": "^0.12.2",
-    "react-tools": "^0.12.2",
+    "react": "^0.13.2",
+    "react-tools": "^0.13.2",
     "underscore": "^1.7.0"
   },
   "devDependencies": {
@@ -42,7 +43,7 @@
     "eslint-plugin-react": "^2.5.0",
     "jsx-loader": "^0.13.2",
     "mocha": "^2.1.0",
-    "npm-shrinkwrap": "^5.4.0",
+    "npm-shrinkwrap": "https://registry.npmjs.org/npm-shrinkwrap/-/npm-shrinkwrap-5.4.0.tgz",
     "sinon": "^1.12.2",
     "source-map-loader": "^0.1.5",
     "transform-loader": "^0.2.2"
