@@ -1,12 +1,10 @@
-/** @jsx React.DOM */
-
 var React = require("react/addons");
 
 var FormGroupComponent = React.createClass({
   displayName: "FormGroupComponent",
   propTypes: {
     errors: React.PropTypes.array,
-    children: React.PropTypes.component.isRequired,
+    children: React.PropTypes.object.isRequired,
     model: React.PropTypes.object.isRequired
   },
 
@@ -40,7 +38,7 @@ var FormGroupComponent = React.createClass({
 
     // Assume there is a single child of either <input> or <textarea>, and add
     // the needed props to make it an input for this attribute.
-    var formControlChild = React.addons.cloneWithProps(
+    var formControlChild = React.cloneElement(
       React.Children.only(this.props.children),
       {
         className: "form-control",
