@@ -1,5 +1,4 @@
-/** @jsx React.DOM */
-
+var classNames = require("classnames");
 var React = require("react/addons");
 
 function roundWorkaround(x) {
@@ -78,7 +77,7 @@ var AppHealthComponent = React.createClass({
       /* jscs:disable disallowTrailingWhitespace, validateQuoteMarks, maximumLineLength */
       return (
         <div
-          className={React.addons.classSet(classSet)}
+          className={classNames(classSet)}
           style={{width: width + "%"}}
           key={i} />
       );
