@@ -1,8 +1,7 @@
-/** @jsx React.DOM */
-
 var $ = require("jquery");
 var _ = require("underscore");
 var React = require("react/addons");
+
 var BackboneMixin = require("../mixins/BackboneMixin");
 var App = require("../models/App");
 var FormGroupComponent = require("../components/FormGroupComponent");
