@@ -1,6 +1,5 @@
-/** @jsx React.DOM */
-
 var React = require("react/addons");
+
 var Info = require("../../models/Info");
 var BackboneMixin = require("../../mixins/BackboneMixin");
 var ModalComponent = require("../components/../ModalComponent");
